+++ backend/src/main/java/com/miu/Book.java
@@ -1,4 +1,4 @@
-package com.miu.model;
+package com.miu;
 import java.io.Serializable;
 import java.util.List;
