+++ backend/src/main/java/com/miu/BookFactory.java
@@ -1,0 +1,4 @@
+package com.miu;
+
+public class BookFactory {
+}
