+++ backend/src/main/java/com/miu/LibraryMember.java
@@ -1,4 +1,4 @@
-package com.miu.model;
+package com.miu;
 
 import java.util.List;
