--- backend/src/main/java/com/miu/controller/UserController.java
@@ -1,10 +1,0 @@
-package com.miu.controller;
-
-public class UserController {
-
-
-    public void add(){
-
-    }
-
-}
