--- backend/src/main/java/com/miu/service/UserService.java
@@ -1,10 +1,0 @@
-package com.miu.service;
-
-import com.miu.model.User;
-
-public interface UserService {
-
-    User login();
-    User findByUsername(String username);
-
-}
